+++ src/GatewayIntegrations.php
@@ -10,6 +10,8 @@
 
 namespace Pronamic\WordPress\Pay;
 
+use Pronamic\WordPress\Pay\Gateways\Common\AbstractIntegration;
+
 /**
  * Title: WordPress gateway integrations class.
  *
@@ -21,7 +23,7 @@ class GatewayIntegrations {
 	/**
 	 * Integrations.
 	 *
-	 * @var array
+	 * @var AbstractIntegration[]
 	 */
 	private $integrations = array();
 
@@ -58,13 +60,15 @@ class GatewayIntegrations {
 			}
 
 			$this->integrations[ $integration->get_id() ] = $integration;
+
+			$this->maybe_add_provider( $integration );
 		}
 	}
 
 	/**
 	 * Register gateway integrations.
 	 *
-	 * @return array
+	 * @return AbstractIntegration[]
 	 */
 	public function register_integrations() {
 		// Register config providers.
@@ -73,5 +77,29 @@ class GatewayIntegrations {
 		}
 
 		return $this->integrations;
+	}
+
+	/**
+	 * Maybe add provider from gateway integration.
+	 *
+	 * @param AbstractIntegration $integration Gateway integration.
+	 */
+	public function maybe_add_provider( $integration ) {
+		global $pronamic_pay_providers;
+
+		if ( ! is_callable( array( $integration, 'get_provider' ) ) ) {
+			return;
+		}
+
+		$provider = $integration->get_provider();
+
+		if ( isset( $pronamic_pay_providers[ $provider ] ) ) {
+			return;
+		}
+
+		$pronamic_pay_providers[ $provider ] = array(
+			'name' => $integration->get_name(),
+			'url'  => $integration->get_url(),
+		);
 	}
 }